upload
This project is of for security system face is recognized and get into the payment mood . And that server will take upon to another on. I have done with flask programming and python .To get access the AWS we had done some thing in uduntu os of my friends. EC2 module is very much help full for doing this one in the instances . Basically this video is of the for presenting the project.